Wed, Aug 17, 2022, 9:26 PM +00:00

+++ credit/credit.c
@@ -12,7 +12,7 @@ int main(void)
     }
     while (credit <= 0);
     //Checksum
-    int answer = checksum(credit);
+    long answer = checksum(credit);
     if (answer % 10 == 0)
     {
         //AMEX
